Week7/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
pyyaml>=6.0.0
google-generativeai>=0.8.0
python-dotenv>=1.0.0
pillow>=10.0.0
